Revert "feat: verify command (#279)"
This reverts commit aed36d2ea7a39ea1ae6bbd5f8d01ef0ce7523cde.
