+++ src/anemoi/datasets/data/__init__.py
@@ -8,7 +8,6 @@
 # nor does it submit to any jurisdiction.
 
 import logging
-import os
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import Set
@@ -83,9 +82,6 @@ def open_dataset(*args: Any, **kwargs: Any) -> "Dataset":
     Dataset
         The opened dataset.
     """
-
-    trace = int(os.environ.get("ANEMOI_DATASETS_TRACE", 0))
-
     # That will get rid of OmegaConf objects
 
     args, kwargs = _convert(args), _convert(kwargs)
@@ -94,12 +90,6 @@ def open_dataset(*args: Any, **kwargs: Any) -> "Dataset":
     ds = ds.mutate()
     ds.arguments = {"args": args, "kwargs": kwargs}
     ds._check()
-
-    if trace:
-        from anemoi.datasets.testing import Trace
-
-        ds = Trace(ds)
-
     return ds
